+++ RELEASES.md
@@ -1,6 +1,7 @@
 Version 0.8.10 (2021-XX-XX)
 ========================
  * RAM 1500 2019-2021 support thanks to the Emergency Midnight RAM Porting Team!
+ * Škoda Kamiq 2021 support thanks to jyoung8607!
  * Škoda Karoq 2019 support thanks to jyoung8607!
 
 Version 0.8.9 (2021-09-14)